Refactor UI elements.
