+++ js/main2e.js
@@ -1,8 +1,6 @@
 import MicroModal from 'micromodal';
 
 import Apple2IO from './apple2io';
-import ApplesoftDump from './applesoft/decompiler';
-import ApplesoftCompiler from './applesoft/compiler';
 import { HiresPage, LoresPage, VideoModes } from './canvas';
 import CPU6502 from './cpu6502';
 import MMU from './mmu';
@@ -10,13 +8,12 @@ import Prefs from './prefs';
 import { debug, gup, hup } from './util';
 
 import Audio from './ui/audio';
-import DriveLights from './ui/drive_lights';
+import { driveLights, initUI, loadAjax, doLoadHTTP } from './ui/apple2';
 import { gamepad, configGamepad, initGamepad, processGamepad } from './ui/gamepad';
 import KeyBoard from './ui/keyboard';
 import Printer from './ui/printer';
-import Tape, { TAPE_TYPES } from './ui/tape';
 
-import DiskII, { DISK_TYPES } from './cards/disk2';
+import DiskII from './cards/disk2';
 import Parallel from './cards/parallel';
 import RAMFactor from './cards/ramfactor';
 import Thunderclock from './cards/thunderclock';
@@ -29,6 +26,12 @@ import Apple2eROM from './roms/apple2e';
 import Apple2eEnhancedROM from './roms/apple2enh';
 
 import SYMBOLS from './symbols';
+
+export {
+    handleDrop, handleDragOver, handleDragEnd,
+    openLoad, openSave,
+    selectCategory, selectDisk, clickDisk, doLoad
+} from './ui/apple2';
 
 var kHz = 1023;
 
@@ -45,15 +48,7 @@ var TRACE = false;
 var MAX_TRACE = 256;
 var trace = [];
 
-var disk_categories = {'Local Saves': []};
-var disk_sets = {};
-var disk_cur_name = [];
-var disk_cur_cat = [];
-
-var tape;
 var disk2;
-var driveLights;
-var _currentDrive = 1;
 
 export function setTrace(debug, trace) {
     DEBUG = debug;
@@ -62,225 +57,6 @@ export function setTrace(debug, trace) {
 
 export function getTrace() {
     return trace;
-}
-
-export function openLoad(drive, event)
-{
-    _currentDrive = parseInt(drive, 10);
-    if (event.metaKey) {
-        openLoadHTTP(drive);
-    } else {
-        if (disk_cur_cat[drive]) {
-            document.querySelector('#category_select').value = disk_cur_cat[drive];
-            selectCategory();
-        }
-        MicroModal.show('load-modal');
-    }
-}
-
-export function openSave(drive, event)
-{
-    _currentDrive = parseInt(drive, 10);
-
-    var mimetype = 'application/octet-stream';
-    var data = disk2.getBinary(drive);
-    var a = document.querySelector('#local_save_link');
-
-    var blob = new Blob([data], { 'type': mimetype });
-    a.href = window.URL.createObjectURL(blob);
-    a.download = driveLights.label(drive) + '.dsk';
-
-    if (event.metaKey) {
-        dumpDisk(drive);
-    } else {
-        document.querySelector('#save_name').value = driveLights.label(drive);
-        MicroModal.show('save-modal');
-    }
-}
-
-export function handleDragOver(drive, event) {
-    event.preventDefault();
-    event.dataTransfer.dropEffect = 'copy';
-}
-
-export function handleDragEnd(drive, event) {
-    var dt = event.dataTransfer;
-    if (dt.items) {
-        for (var i = 0; i < dt.items.length; i++) {
-            dt.items.remove(i);
-        }
-    } else {
-        event.dataTransfer.clearData();
-    }
-}
-
-export function handleDrop(drive, event) {
-    event.preventDefault();
-    event.stopPropagation();
-
-    if (drive < 1) {
-        if (!disk2.getMetadata(1)) {
-            drive = 1;
-        } else if (!disk2.getMetadata(2)) {
-            drive = 2;
-        } else {
-            drive = 1;
-        }
-    }
-    var dt = event.dataTransfer;
-    if (dt.files.length == 1) {
-        doLoadLocal(drive, dt.files[0]);
-    } else if (dt.files.length == 2) {
-        doLoadLocal(1, dt.files[0]);
-        doLoadLocal(2, dt.files[1]);
-    } else {
-        for (var idx = 0; idx < dt.items.length; idx++) {
-            if (dt.items[idx].type === 'text/uri-list') {
-                dt.items[idx].getAsString(function(url) {
-                    var parts = document.location.hash.split('|');
-                    parts[drive - 1] = url;
-                    document.location.hash = parts.join('|');
-                });
-            }
-        }
-    }
-}
-
-var loading = false;
-
-function loadAjax(drive, url) {
-    loading = true;
-    MicroModal.show('loading-modal');
-
-    fetch(url).then(function(response) {
-        return response.json();
-    }).then(function(data) {
-        if (data.type == 'binary') {
-            loadBinary(drive, data);
-        } else if (DISK_TYPES.indexOf(data.type) > -1) {
-            loadDisk(drive, data);
-        }
-        initGamepad(data.gamepad);
-        MicroModal.close('loading-modal');
-        loading = false;
-    }).catch(function(error) {
-        window.alert(error || status);
-        MicroModal.close('loading-modal');
-        loading = false;
-    });
-}
-
-export function doLoad() {
-    MicroModal.close('load-modal');
-    var urls = document.querySelector('#disk_select').value, url;
-    if (urls && urls.length) {
-        if (typeof(urls) == 'string') {
-            url = urls;
-        } else {
-            url = urls[0];
-        }
-    }
-
-    var files = document.querySelector('#local_file').files;
-    if (files.length == 1) {
-        doLoadLocal(_currentDrive, files[0]);
-    } else if (url) {
-        var filename;
-        MicroModal.close('load-modal');
-        if (url.substr(0,6) == 'local:') {
-            filename = url.substr(6);
-            if (filename == '__manage') {
-                openManage();
-            } else {
-                loadLocalStorage(_currentDrive, filename);
-            }
-        } else {
-            var r1 = /json\/disks\/(.*).json$/.exec(url);
-            if (r1) {
-                filename = r1[1];
-            } else {
-                filename = url;
-            }
-            var parts = document.location.hash.split('|');
-            parts[_currentDrive - 1] = filename;
-            document.location.hash = parts.join('|');
-        }
-    }
-}
-
-export function doSave() {
-    var name = document.querySelector('#save_name').value;
-    saveLocalStorage(_currentDrive, name);
-    MicroModal.close('save-modal');
-}
-
-export function doDelete(name) {
-    if (window.confirm('Delete ' + name + '?')) {
-        deleteLocalStorage(name);
-    }
-}
-
-function doLoadLocal(drive, file) {
-    var parts = file.name.split('.');
-    var ext = parts[parts.length - 1].toLowerCase();
-    if (DISK_TYPES.indexOf(ext) > -1) {
-        doLoadLocalDisk(drive, file);
-    } else if (TAPE_TYPES.indexOf(ext) > -1) {
-        tape.doLoadLocalTape(file);
-    } else {
-        window.alert('Unknown file type: ' + ext);
-    }
-}
-
-function doLoadLocalDisk(drive, file) {
-    var fileReader = new FileReader();
-    fileReader.onload = function() {
-        var parts = file.name.split('.');
-        var ext = parts.pop().toLowerCase();
-        var name = parts.join('.');
-        if (disk2.setBinary(drive, name, ext, this.result)) {
-            driveLights.label(drive, name);
-            initGamepad();
-        }
-    };
-    fileReader.readAsArrayBuffer(file);
-}
-
-function doLoadHTTP(drive, _url) {
-    var url = _url || document.querySelector('#http_url').value;
-    if (url) {
-        var req = new XMLHttpRequest();
-        req.open('GET', url, true);
-        req.responseType = 'arraybuffer';
-
-        req.onload = function() {
-            if (req.status !== 200) {
-                return window.alert('Unable to load "' + url + '"');
-            }
-            var urlParts = url.split('/');
-            var file = urlParts.pop();
-            var fileParts = file.split('.');
-            var ext = fileParts.pop().toLowerCase();
-            var name = decodeURIComponent(fileParts.join('.'));
-            if (disk2.setBinary(drive, name, ext, req.response)) {
-                driveLights.label(drive, name);
-                if (!_url) {
-                    MicroModal.close('http-modal');
-                }
-                initGamepad();
-            }
-        };
-        req.send(null);
-    }
-}
-
-function openLoadHTTP(drive) {
-    _currentDrive = parseInt(drive, 10);
-    MicroModal.show('http-modal');
-}
-
-function openManage() {
-    MicroModal.show('manage-modal');
 }
 
 var prefs = new Prefs();
@@ -339,14 +115,10 @@ var hgr2 = new HiresPage(2, context4);
 var vm = new VideoModes(gr, hgr, gr2, hgr2, true);
 vm.enhanced(enhanced);
 vm.multiScreen(multiScreen);
-var dumper = new ApplesoftDump(cpu);
-var compiler = new ApplesoftCompiler(cpu);
 
-driveLights = new DriveLights();
 var io = new Apple2IO(cpu, vm);
 var keyboard = new KeyBoard(cpu, io, true);
 var audio = new Audio(io);
-tape = new Tape(io);
 var printer = new Printer('#printer-modal .paper');
 
 var mmu = new MMU(cpu, vm, gr, gr2, hgr, hgr2, io, rom);
@@ -357,6 +129,8 @@ var parallel = new Parallel(io, 1, printer);
 var slinky = new RAMFactor(io, 2, 1024 * 1024);
 disk2 = new DiskII(io, 6, driveLights);
 var clock = new Thunderclock(io, 7);
+
+initUI(cpu, io, disk2);
 
 io.setSlot(1, parallel);
 io.setSlot(2, slinky);
@@ -401,23 +175,6 @@ export function updateSound() {
         label.classList.remove('fa-volume-up');
         label.classList.add('fa-volume-off');
     }
-}
-
-function dumpDisk(drive) {
-    var wind = window.open('', '_blank');
-    wind.document.title = driveLights.label(drive);
-    wind.document.write('<pre>');
-    wind.document.write(disk2.getJSON(drive, true));
-    wind.document.write('</pre>');
-    wind.document.close();
-}
-
-export function dumpProgram() {
-    debug(dumper.toString());
-}
-
-export function compileProgram(program) {
-    compiler.compile(program);
 }
 
 export function step()
@@ -479,29 +236,27 @@ function run(pc) {
                 processHash(hash);
             }
         }
-        if (!loading) {
-            mmu.resetVB();
-            if (DEBUG) {
-                cpu.stepCyclesDebug(TRACE ? 1 : step, function() {
-                    var line = cpu.dumpRegisters() + ' ' +
-                        cpu.dumpPC(undefined, SYMBOLS);
-                    if (TRACE) {
-                        debug(line);
-                    } else {
-                        trace.push(line);
-                        if (trace.length > MAX_TRACE) {
-                            trace.shift();
-                        }
+        mmu.resetVB();
+        if (DEBUG) {
+            cpu.stepCyclesDebug(TRACE ? 1 : step, function() {
+                var line = cpu.dumpRegisters() + ' ' +
+                    cpu.dumpPC(undefined, SYMBOLS);
+                if (TRACE) {
+                    debug(line);
+                } else {
+                    trace.push(line);
+                    if (trace.length > MAX_TRACE) {
+                        trace.shift();
                     }
-                });
-            } else {
-                cpu.stepCycles(step);
-            }
-            if (vm.blit()) {
-                renderedFrames++;
-            }
-            io.tick();
+                }
+            });
+        } else {
+            cpu.stepCycles(step);
         }
+        if (vm.blit()) {
+            renderedFrames++;
+        }
+        io.tick();
 
         processGamepad(io);
 
@@ -578,130 +333,6 @@ function restoreState() {
     driveLights.setState(state.driveLights);
     if (slinky && state.slinky) {
         slinky.setState(state.slinky);
-    }
-}
-
-function loadBinary(bin) {
-    stop();
-    for (var idx = 0; idx < bin.length; idx++) {
-        var pos = bin.start + idx;
-        cpu.write(pos >> 8, pos & 0xff, bin.data[idx]);
-    }
-    run(bin.start);
-}
-
-export function selectCategory() {
-    document.querySelector('#disk_select').innerHTML = '';
-    var cat = disk_categories[document.querySelector('#category_select').value];
-    if (cat) {
-        for (var idx = 0; idx < cat.length; idx++) {
-            var file = cat[idx], name = file.name;
-            if (file.disk) {
-                name += ' - ' + file.disk;
-            }
-            var option = document.createElement('option');
-            option.value = file.filename;
-            option.innerText = name;
-            document.querySelector('#disk_select').append(option);
-            if (disk_cur_name[_currentDrive] == name) {
-                option.selected = true;
-            }
-        }
-    }
-}
-
-export function selectDisk() {
-    document.querySelector('#local_file').value = '';
-}
-
-export function clickDisk() {
-    doLoad();
-}
-
-function loadDisk(drive, disk) {
-    var name = disk.name;
-    var category = disk.category;
-
-    if (disk.disk) {
-        name += ' - ' + disk.disk;
-    }
-
-    disk_cur_cat[drive] = category;
-    disk_cur_name[drive] = name;
-
-    driveLights.label(drive, name);
-    disk2.setDisk(drive, disk);
-    initGamepad(disk.gamepad);
-}
-
-/*
- *  LocalStorage Disk Storage
- */
-
-function updateLocalStorage() {
-    var diskIndex = JSON.parse(window.localStorage.diskIndex || '{}');
-    var names = [], name, cat;
-
-    for (name in diskIndex) {
-        if (diskIndex.hasOwnProperty(name)) {
-            names.push(name);
-        }
-    }
-
-    cat = disk_categories['Local Saves'] = [];
-    document.querySelector('#manage-modal-content').innerHTML = '';
-
-    names.forEach(function(name) {
-        cat.push({
-            'category': 'Local Saves',
-            'name': name,
-            'filename': 'local:' + name
-        });
-        document.querySelector('#manage-modal-content').innerHTML =
-            '<span class="local_save">' +
-            name +
-            ' <a href="#" onclick="doDelete(\'' +
-            name +
-            '\')">Delete</a><br /></span>';
-    });
-    cat.push({
-        'category': 'Local Saves',
-        'name': 'Manage Saves...',
-        'filename': 'local:__manage'
-    });
-}
-
-function saveLocalStorage(drive, name) {
-    var diskIndex = JSON.parse(window.localStorage.diskIndex || '{}');
-
-    var json = disk2.getJSON(drive);
-    diskIndex[name] = json;
-
-    window.localStorage.diskIndex = JSON.stringify(diskIndex);
-
-    window.alert('Saved');
-
-    driveLights.label(drive, name);
-    driveLights.dirty(drive, false);
-    updateLocalStorage();
-}
-
-function deleteLocalStorage(name) {
-    var diskIndex = JSON.parse(window.localStorage.diskIndex || '{}');
-    if (diskIndex[name]) {
-        delete diskIndex[name];
-        window.alert('Deleted');
-    }
-    window.localStorage.diskIndex = JSON.stringify(diskIndex);
-    updateLocalStorage();
-}
-
-function loadLocalStorage(drive, name) {
-    var diskIndex = JSON.parse(window.localStorage.diskIndex || '{}');
-    if (diskIndex[name]) {
-        disk2.setJSON(drive, diskIndex[name]);
-        driveLights.label(drive, name);
-        driveLights.dirty(drive, false);
     }
 }
 
@@ -943,39 +574,6 @@ document.addEventListener('DOMContentLoaded', function() {
     updateSound();
     updateScreen();
     updateCPU();
-
-    if (window.localStorage !== undefined) {
-        document.querySelectorAll('.disksave').forEach(function (el) { el.style.display = 'inline-block';});
-    }
-
-    var oldcat = '';
-    var option;
-    for (var idx = 0; idx < window.disk_index.length; idx++) {
-        var file = window.disk_index[idx];
-        var cat = file.category;
-        var name = file.name, disk = file.disk;
-        if (cat != oldcat) {
-            option = document.createElement('option');
-            option.value = cat;
-            option.innerText = cat;
-            document.querySelector('#category_select').append(option);
-
-            disk_categories[cat] = [];
-            oldcat = cat;
-        }
-        disk_categories[cat].push(file);
-        if (disk) {
-            if (!disk_sets[name]) {
-                disk_sets[name] = [];
-            }
-            disk_sets[name].push(file);
-        }
-    }
-    option = document.createElement('option');
-    option.innerText = 'Local Saves';
-    document.querySelector('#category_select').append(option);
-
-    updateLocalStorage();
     initGamepad();
 
     // Check for disks in hashtag
